Corrections to ratelimit calculations
Corrections to ratelimit calculations to ensure correct timing according
the the rate limit hertz configuration setting.  Previous version had a
bug in the calculations.  Also change quadrature output to default to 0
when idle.  This is to prevent SmallyMouse2 from supplying power to the
host (when SmallyMouse is self-powered and the host is off).